Initial progress in 64-bit lame mp3 encoder.  Upgraded the project to VS2010.

+++ ACM/main.cpp
@@ -35,7 +35,7 @@
 /// So the new/delete operators have to be overriden in order to use memory
 /// readable out of the calling process
 
-void * operator new( unsigned int cb )
+void * operator new( size_t cb )
 {
 	return LocalAlloc(LPTR, cb); // VirtualAlloc
 }
